Update README.md
This project is a Weather App that provides real-time weather information and a 5-day forecast for a given city. The app features a dynamic video background for an immersive user experience. Users can enter a city name, and the app will fetch and display the current weather (temperature, description, humidity, wind speed) along with weather emojis representing the forecast.